Refactored serial/RTU to match TCP
Refactored the sync and async RTU classes to match the TCP counterparts. Instead of `Serial` being both a modbus RTU client *and* server, the server functionality has been split off into the `RTUServer` class, and `Serial` is now only the RTU client. `ModbusRTU` remains unaffected as its `_itf` is changed from `Serial` to `RTUServer`, but the user-facing API remains the same as they handle the same tasks as before.